[tools] Rewrite `et ci-status` as interactive `et ci-inspect` TUI
Replace the static table-based CI metrics output with an interactive
terminal UI matching the `et github-inspect` pattern. Browse workflow
health categories, drill into individual workflows for daily trends
and failure details, and inspect error logs inline.
